Migrate from JSON parsing to Viper for configuration. Solves #29
The New() function to create a bot now uses Viper to load the config
The config file can be provided as before. If no config file is provided
Slick defaults to a local config.json or alternatively a config.json in
$HOME/.slick.

The old Loadconfig() function now uses Vipers unmarshalling capability.
Config is only read once.